pyproject.toml
[project]
name = "flaretorch"
version = "0.1.0"
description = "Machine learning library for solar flare forecasting with uncertainty quantification"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jinsu Hong", email = "jinsuhong.knight@gmail.com"},
    {name = "Berkay Aydin", email = "baydin2@gsu.edu"},
]
maintainers = [
    {name = "Jinsu Hong", email = "jinsuhong.knight@gmail.com"},
]
keywords = [
    "deep-learning",
    "pytorch",
    "solar-flare",
    "uncertainty-quantification",
    "machine-learning",
    "forecasting",
    "pytorch-lightning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "curvlinops-for-pytorch<3.0",
    "dask>=2025.11.0",
    "datasets>=4.4.1",
    "hydra-core>=1.3.2",
    "laplace-torch>=0.2.2.2",
    "loguru>=0.7.3",
    "terratorch>=1.1.1",
    "terratorch-surya>=0.1.0",
    "transformers>=4.36.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.2",
    "mypy>=1.8",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=2.0",
]

[project.urls]
Homepage = "https://github.com/JinsuHongg/FlareTorch"
Documentation = "https://flaretorch.readthedocs.io"
Repository = "https://github.com/JinsuHongg/FlareTorch.git"
Issues = "https://github.com/JinsuHongg/FlareTorch/issues"
Changelog = "https://github.com/JinsuHongg/FlareTorch/blob/main/CHANGELOG.md"

[project.scripts]
flare-train = "flaretorch.cli.train:main"
flare-eval = "flaretorch.cli.eval:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/configs",
]

[tool.hatch.build.targets.wheel]
packages = ["src/flaretorch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
